@@ -1,7 +1,7 @@
 import { answerTask, getCampaignProgress, getTask } from "api/reviews";
 import { InformationModal } from "components/Error";
-import React, { useEffect, useState } from "react";
-import { Redirect, useParams } from "react-router-dom";
+import { useEffect, useState } from "react";
+import { Navigate, useParams } from "react-router-dom";
 import {
   Button,
   Grid,
@@ -492,7 +492,10 @@ function CampaignTask() {
     }
     fetchTask();
     // eslint-disable-next-line no-return-assign
-    return () => (isCancelled = true);
+    return function ignoreFetchResults() {
+      isCancelled = true;
+      return null;
+    };
   }, [id, mode, tasksDone]);
 
   useEffect(() => {
@@ -503,10 +506,11 @@ function CampaignTask() {
   }, [id, mode]); // We do not re-render this, just so that we don't spam the server.
 
   if (tasksDone >= tasksTotal) {
-    return <Redirect to="/campaigns" />;
+    return <Navigate to="/campaigns" />;
   }
 
   if (task === null) {
+    // eslint-disable-next-line
     return <></>;
   }
 
