@@ -3,7 +3,12 @@ module.exports = {
     browser: true,
     es6: true,
   },
-  extends: ["plugin:react/recommended", "airbnb", "prettier"],
+  extends: [
+    "airbnb",
+    "plugin:react/recommended",
+    "plugin:react/jsx-runtime",
+    "plugin:prettier/recommended",
+  ],
   globals: {
     Atomics: "readonly",
     SharedArrayBuffer: "readonly",
@@ -16,15 +21,16 @@ module.exports = {
     ecmaVersion: 2018,
     sourceType: "module",
   },
-  plugins: ["react", "prettier"],
+  plugins: ["react", "eslint-plugin-react-hooks", "prettier"],
   rules: {
-    "prettier/prettier": "error",
     "react/jsx-filename-extension": [1, { extensions: [".js", ".jsx"] }],
     "react/prop-types": [0],
     "react/destructuring-assignment": [0],
+    "react-hooks/rules-of-hooks": "error", // Checks rules of Hooks
+    "react-hooks/exhaustive-deps": "warn", // Checks effect dependencies
     "no-console": [0],
     "no-unused-vars": ["error", { argsIgnorePattern: "^_" }],
-    "import/no-extraneous-dependencies": ["error", { devDependencies: true }]
+    "import/no-extraneous-dependencies": ["error", { devDependencies: true }],
   },
   settings: {
     "import/resolver": {
