@@ -3,7 +3,7 @@ import {
   setTargetLanguage,
   swapLanguages,
 } from "features/translate/translateSlice";
-import React, { useEffect, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useDispatch, useSelector } from "react-redux";
 import { Button, Dropdown } from "semantic-ui-react";
@@ -26,44 +26,48 @@ function LanguagePicker(props) {
   );
 }
 
-const LanguageDirectionPicker = () => {
+function LanguageDirectionPicker() {
   const { t } = useTranslation();
   const { sourceLang, targetLang } = useSelector((state) => state.translation);
   const [tgtLanguageOptions, setTgtLanuageOptions] = useState([]);
   const dispatch = useDispatch();
   // TODO: Get this information from the backend
-  const SUPPORTED_LANGUAGE_PAIRS = {
-    en: ["is", "fo"],
-    is: ["en", "pl"],
-    pl: ["is"],
-    fo: ["en"],
-  };
-  const LANGUAGE_OPTIONS = {
-    is: {
-      key: "is",
-      text: t("Icelandic"),
-      value: "is",
-      flag: "is",
-    },
-    en: {
-      key: "en",
-      text: t("English"),
-      value: "en",
-      flag: "uk",
-    },
-    pl: {
-      key: "pl",
-      text: t("Polish"),
-      value: "pl",
-      flag: "pl",
-    },
-    fo: {
-      key: "fo",
-      text: t("Faroese"),
-      value: "fo",
-      flag: "fo",
-    },
-  };
+  const SUPPORTED_LANGUAGE_PAIRS = useMemo(() => {
+    return {
+      en: ["is", "fo"],
+      is: ["en", "pl"],
+      pl: ["is"],
+      fo: ["en"],
+    };
+  }, []);
+  const LANGUAGE_OPTIONS = useMemo(() => {
+    return {
+      is: {
+        key: "is",
+        text: t("Icelandic"),
+        value: "is",
+        flag: "is",
+      },
+      en: {
+        key: "en",
+        text: t("English"),
+        value: "en",
+        flag: "uk",
+      },
+      pl: {
+        key: "pl",
+        text: t("Polish"),
+        value: "pl",
+        flag: "pl",
+      },
+      fo: {
+        key: "fo",
+        text: t("Faroese"),
+        value: "fo",
+        flag: "fo",
+      },
+    };
+  }, [t]);
   useEffect(() => {
     // TODO: Consider moving valid target lang logic to redux
     const validTgtLanguages = SUPPORTED_LANGUAGE_PAIRS[sourceLang];
@@ -73,7 +77,13 @@ const LanguageDirectionPicker = () => {
       const newTargetLang = validTgtLanguages[0];
       dispatch(setTargetLanguage(newTargetLang));
     }
-  }, [sourceLang]);
+  }, [
+    sourceLang,
+    LANGUAGE_OPTIONS,
+    SUPPORTED_LANGUAGE_PAIRS,
+    dispatch,
+    targetLang,
+  ]);
   const SRC_LANGUAGE_OPTIONS = Object.keys(LANGUAGE_OPTIONS).map(
     (key) => LANGUAGE_OPTIONS[key]
   );
@@ -102,6 +112,6 @@ const LanguageDirectionPicker = () => {
       </div>
     </div>
   );
-};
+}
 
 export default LanguageDirectionPicker;
