@@ -1,18 +1,40 @@
-import { configureStore, getDefaultMiddleware } from "@reduxjs/toolkit";
+import { combineReducers, configureStore } from "@reduxjs/toolkit";
+import { languageSettingReducer } from "features/i18n/languageSettingSlice";
 import { loginReducer } from "features/login/loginSlice";
 import { translateReducer } from "features/translate/translateSlice";
 import logger from "redux-logger";
+import {
+  FLUSH,
+  PAUSE,
+  PERSIST,
+  persistReducer,
+  persistStore,
+  PURGE,
+  REGISTER,
+  REHYDRATE,
+} from "redux-persist";
+import storage from "redux-persist/lib/storage"; // defaults to localStorage for web
 
-const reducer = {
+const rootReducer = combineReducers({
   login: loginReducer,
   translation: translateReducer,
-};
-
-const middleware = [...getDefaultMiddleware(), logger];
+  language: languageSettingReducer,
+});
 
 // eslint-disable-next-line import/prefer-default-export
+const persistConfig = {
+  key: "root",
+  storage,
+};
+const persistedReducer = persistReducer(persistConfig, rootReducer);
 export const store = configureStore({
-  reducer,
-  middleware,
+  reducer: persistedReducer,
+  middleware: (getDefaultMiddleware) =>
+    getDefaultMiddleware({
+      serializableCheck: {
+        ignoredActions: [FLUSH, REHYDRATE, PAUSE, PERSIST, PURGE, REGISTER],
+      },
+    }).concat(logger),
   devTools: process.env.NODE_ENV !== "production",
 });
+export const persistor = persistStore(store);
